requirements.txt
anyio==3.7.1
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.4
djangorestframework==3.14.0
exceptiongroup==1.1.3
flake8==6.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
mccabe==0.7.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
